fix: rename SelectFix to SelectInline
