+++ src/components/index.ts
@@ -20,6 +20,7 @@ export { default as RobustNotice } from './Notice/Notice.vue'
 export { default as RobustPopper } from './Popper/Popper'
 export { default as RobustRadio } from './Radio/Radio.vue'
 export { default as RobustSelect } from './Select/Select.vue'
+export { default as RobustSelectInline } from './SelectInline/SelectInline.vue'
 export { default as RobustSeparator } from './Separator/Separator.vue'
 export { default as RobustSlider } from './Slider/Slider.vue'
 export { default as RobustSpinner } from './Spinner/Spinner.vue'
